refactor: unify Memoria client — CLI delegates to astra-tools
CLI edge_tools/memoria.rs (784L → 451L) now delegates to the shared
astra_tools::memoria::MemoriaClient for all dispatch logic:

Moved to shared client:
- HttpMethod enum (Get/Put/Post)
- build_direct_request with correct method routing (GET for profile,
  PUT for correct-by-id, POST for rest)
- Type normalization (business types → V1 primitives) in call path
- Cloud helpers (snapshot/branch/reflect/health) via memoria_cloud_request
- memoria_oneshot_client (was duplicated)
- memoria_governance/consolidate_fire_and_forget

Fixed in shared client:
- /v1/memories/governance → /v1/governance (matched actual Memoria API)
- /v1/memories/consolidate → /v1/consolidate (same fix)
- correct endpoint routing: memory_id → PUT, query → POST

CLI retains only:
- Circuit breaker with user notification (⚠/✓ messages)
- memory_boost_search (CLI display logic)
- memory_feedback_useful (CLI lifecycle)
- memoria_retrieve_lessons (session bootstrap)
- memoria_store_lessons_fire_and_forget (session-end)
